Move NodeDeletionTracker to a separate package
